Adapt Parser tests to ComponentPreprocessor
